+++ resources/views/components/persistent.blade.php
@@ -1,0 +1,14 @@
+@props([
+    'id' => null,
+])
+
+@php
+    $persistentId = $id ?? 'persistent-' . \Illuminate\Support\Str::random(8);
+@endphp
+
+<div
+    data-accelade-persistent="{{ $persistentId }}"
+    {{ $attributes }}
+>
+    {{ $slot }}
+</div>